Move ExecutablePexTool to the python backend.
It depends on code from this backend, which is
inappropriate for something in the core.

In any case, it was only used in one place, back in the
backend, so there's no need for it to be in the core anyway.
