+++ src/python/pants/binaries/BUILD
@@ -5,7 +5,6 @@ python_library(
   dependencies=[
     '3rdparty/python/twitter/commons:twitter.common.collections',
     '3rdparty/python:dataclasses',
-    '3rdparty/python:pex',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/engine:fs',
